at: Allow for v1.1.1~rc1 version numbering
Tags in git cannot have a tilde in the tag name.

Source tarball versioning for deb and rpm package creation needs to use
a tilde AOT a '-' as the '~' is sorted before the plain version.
i.e. 1.1.1~rc1 < 1.1.1 < 1.1.1-rc1

So a release tagged at -rc1 must be renamed to ~rc1

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>